Refactor: move basic test source into test/src/
- This commit refactor the test suite by moving the basic functional and
  test source files into the test/src/ directory.

- The changes in this commit are:
  - The main test list (ALL_TESTS) in test/mk/components.mk is broken
    down into logical groups (BASIC_TESTS, ACVP_TESTS, BENCH_TESTS).
  - The build logic is updated to locate the basic tests in test/src/,
    while still finding acvp and bench tests in their original location.
  - The include paths within the moved source files
    (gen_KAT.c, test_mldsa.c, test_unit.c, test_alloc.c) are updated to
    reflect their new related path.
  - This commit also prepares for subsequent changes to move
    benchmarking and ACVP test sources into their own dedicated
    directories(acvp, bench).
